Separate manageYAML and Snapcraft in diff files
